.h and .cpp structure
