+++ CMakeLists.txt
@@ -14,11 +14,12 @@ add_subdirectory("dependencies/glfw")
 # Compile GLAD
 add_library(glad dependencies/glad/src/gl.c)
 
+add_executable(Renderden Renderden/src/main.cpp Renderden/src/renderden/Renderden.cpp Renderden/src/renderden/Renderden.h Renderden/src/shapes/Shape.cpp Renderden/src/shapes/Shape.h)
+
+target_link_libraries(Renderden glfw glad)
+
 target_include_directories(glad PUBLIC dependencies/glad/include)
 include_directories(dependencies)
-
-# Specify project directory
-add_subdirectory("Renderden")
 
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
 set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
