--- Renderden/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_executable(Renderden src/main.cpp)
-
-target_link_libraries(Renderden glfw glad)
