+++ Renderden/src/main.cpp
@@ -1,8 +1,9 @@
-#include "Renderden.cpp"
+#include "renderden/Renderden.h"
 
 using namespace Renderden;
 
 int main() {
     RenderdenEngine rd;
     rd.RenderdenInitFunction();
+    std::cout << "Hello";
 }
