+++ Renderden/src/shapes/Shape.cpp
@@ -4,15 +4,12 @@
 #include <stdio.h>
 #include <iostream>
 #include <vector>
-#include "../Renderden.cpp"
-#include "../shaders/Shader.h"
 
-using namespace Renderden;
+#include "../Renderden/src/shapes/Shape.h"
 
-class Shape: public RenderdenEngine {
-    public:
-        int drawCall(std::vector<int> VertexArray) { 
-            std::cout << "Draw call called" << std::endl;
-        }
-};
-
+int Renderden::Shape::drawCall(std::vector<int> VertexArray)
+{
+    std::cout << "Draw call called" << std::endl;
+    // std::cout << this.name;
+    return 1;
+}
